🧪 9E8DFM test: split no-op close regression
