+++ .agentplane/tasks/202606050702-9E8DFM/pr/diffstat.txt
@@ -1,7 +1,7 @@
  .agentplane/tasks/202606040927-KSESDS/README.md    | 116 ++++++++++++++++++++
  .agentplane/tasks/202606041702-TVTSM2/README.md    | 120 +++++++++++++++++++++
- .../src/cli/run-cli.core.tasks.lifecycle.test.ts   | 101 +++++++++++++++++
+ .../run-cli.core.tasks.close-noop-readme.test.ts   | 120 +++++++++++++++++++++
  .../src/commands/task/close-duplicate.ts           |  59 +---------
  .../agentplane/src/commands/task/close-noop.ts     |   3 +-
  .../agentplane/src/commands/task/close-shared.ts   |  57 +++++++++-
- 6 files changed, 396 insertions(+), 60 deletions(-)
+ 6 files changed, 415 insertions(+), 60 deletions(-)
